+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.3.33] - 2019-02-17
 ### Added
 - Config to disable heart color looping [#155]
+- Config to scale player bonus regen with max health [#147]
 ### Fixed
 - Crash when connecting to servers in some cases [#157]
 - Heart container drops not respecting doMobLoot game rule [#156]